+++ src/app/(auth)/mypage/page.tsx
@@ -1,7 +1,7 @@
-import ActiveStudies from "@/features/mypage/components/ActiveStudies";
-import ProfileCard from "@/features/mypage/components/ProfileCard";
-import Statistics from "@/features/mypage/components/Statistics";
-import StudySchedule from "@/features/mypage/components/StudySchedule";
+import ActiveStudies from "@/widget/mypage/ui/ActiveStudies";
+import ProfileCard from "@/widget/mypage/ui/ProfileCard";
+import Statistics from "@/widget/mypage/ui/Statistics";
+import StudySchedule from "@/widget/mypage/ui/StudySchedule";
 
 const MyPage = async () => {
   return (
