+++ src/widget/mypage/ui/StudySchedule.tsx
